Molar eq input improved
Since only one value needs to be entered for molar equivalents (only one can be in excess while the other should be 1.0) and a 1:1 should not require values to be manually entered, molar eq input now works like an input field that can be toggled on and off. One known issue is that strings are accepted as inputs.